+++ components/index.js
@@ -1,24 +1,24 @@
 (function (global, factory) {
   if (typeof define === "function" && define.amd) {
-    define(['exports', './Canvas', './Link', './Node', './Inspector', './Selector'], factory);
+    define(['exports', './Frame', './Link', './Node', './Inspector', './Selector'], factory);
   } else if (typeof exports !== "undefined") {
-    factory(exports, require('./Canvas'), require('./Link'), require('./Node'), require('./Inspector'), require('./Selector'));
+    factory(exports, require('./Frame'), require('./Link'), require('./Node'), require('./Inspector'), require('./Selector'));
   } else {
     var mod = {
       exports: {}
     };
-    factory(mod.exports, global.Canvas, global.Link, global.Node, global.Inspector, global.Selector);
+    factory(mod.exports, global.Frame, global.Link, global.Node, global.Inspector, global.Selector);
     global.index = mod.exports;
   }
-})(this, function (exports, _Canvas, _Link, _Node, _Inspector, _Selector) {
+})(this, function (exports, _Frame, _Link, _Node, _Inspector, _Selector) {
   'use strict';
 
   Object.defineProperty(exports, "__esModule", {
     value: true
   });
-  exports.Selector = exports.Node = exports.Link = exports.Inspector = exports.Canvas = undefined;
+  exports.Selector = exports.Node = exports.Link = exports.Inspector = exports.Frame = undefined;
 
-  var _Canvas2 = _interopRequireDefault(_Canvas);
+  var _Frame2 = _interopRequireDefault(_Frame);
 
   var _Link2 = _interopRequireDefault(_Link);
 
@@ -34,7 +34,7 @@
     };
   }
 
-  exports.Canvas = _Canvas2.default;
+  exports.Frame = _Frame2.default;
   exports.Inspector = _Inspector2.default;
   exports.Link = _Link2.default;
   exports.Node = _Node2.default;